refactor: reorganize agent workflow imports and implement Sub37 agent workflow
